backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
httpx==0.28.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.14.1
cryptography==44.0.0
python-multipart==0.0.20
email-validator==2.3.0
bcrypt==4.1.3
